add page falut trap

+++ trap.c
@@ -14,6 +14,8 @@ extern uint vectors[];  // in vectors.S: array of 256 entry pointers
 struct spinlock tickslock;
 uint ticks;
 
+int mappages(pde_t *pgdir, void *va, uint size, uint pa, int perm);  
+
 void
 tvinit(void)
 {
@@ -86,6 +88,15 @@ trap(struct trapframe *tf)
               tf->trapno, cpuid(), tf->eip, rcr2());
       panic("trap");
     }
+
+	if (tf->trapno == T_PGFLT) {
+		cprintf("page fault\n");
+		myproc()->killed = 1;
+		void *va = (void *)PGROUNDDOWN(rcr2());
+		mappages(myproc()->pgdir, (char *)va, PGSIZE, V2P(kalloc()), 6);
+		break;
+	}
+
     // In user space, assume process misbehaved.
     cprintf("pid %d %s: trap %d err %d on cpu %d "
             "eip 0x%x addr 0x%x--kill proc\n",
